aws-backend/sgmkr-ep-embed-image/requirements.txt
transformers[torch]
pillow
